udpate payment config
