+++ .env
@@ -1,9 +1,31 @@
-FOLDER="./result"
-LIVEBLOCKS_SECRET_KEY=""
-GOOGLE_CLIENT_ID=""
-GOOGLE_CLIENT_SECRET=""
-GOOGLE_REDIRECT_URI="http://localhost:3000/api/auth/callback/google"
-NEXTAUTH_SECRET=""
-NEXTAUTH_URL="http://localhost:3000"
-JWT_SECRET=""
-DATABASE_URL=""
+# 輸出資料夾
+FOLDER=./result
+# Google OAuth 2.0 Client ID
+GOOGLE_CLIENT_ID=
+# Google OAuth 2.0 Client Secret
+GOOGLE_CLIENT_SECRET=
+# Google OAuth 2.0 Redirect URI
+GOOGLE_REDIRECT_URI=http://localhost:3000/api/auth/callback/google
+# NextAuth.js Secret
+NEXTAUTH_SECRET=
+# NextAuth.js URL
+NEXTAUTH_URL=http://localhost:3000
+# JWT 密鑰
+JWT_SECRET=
+# 資料庫連線 URL
+DATABASE_URL=postgresql://account:password@127.0.0.1:5432/database
+
+# 商店 ID
+MERCHANT_ID=
+# 商店 HashKey
+HASHKEY=
+# 商店 HashIV
+HASHIV=
+# 串接程式版本，2023 年版本為 2.0
+PAYMENT_VERSION=2.0
+# 支付完成後，用戶轉址
+PAYMENT_RETURN_URL=http://localhost:3000/
+# 支付完成後，回傳商店網址
+PAYMENT_NOTIFY_URL=http://localhost:3000/api/payment/callback
+# 測試機、正式機 API 路徑
+PAYMENT_URL=https://ccore.newebpay.com/MPG/mpg_gateway
