+++ package.json
@@ -9,6 +9,7 @@
   },
   "dependencies": {
     "@prisma/client": "^5.5.2",
+    "dotenv": "^16.3.2",
     "next": "latest",
     "react": "18.2.0",
     "react-dom": "18.2.0",
